fix: resolve implicit casting issue in compound assignment operations

+++ src/main/java/oscar/form/graphic/FrmPdfGraphicRourke.java
@@ -153,8 +153,8 @@ public class FrmPdfGraphicRourke extends FrmPdfGraphic {
 
             MiscUtils.getLogger().debug("sday, eday " + sday + ", " + eday);
 
-            smonth += (sday / startDate.getActualMaximum(Calendar.DAY_OF_MONTH));
-            emonth += (eday / curDate.getActualMaximum(Calendar.DAY_OF_MONTH));
+            smonth += (sday / (float) startDate.getActualMaximum(Calendar.DAY_OF_MONTH));
+            emonth += (eday / (float) curDate.getActualMaximum(Calendar.DAY_OF_MONTH));
 
             //don't forget to add years
             switch (xDateScale) {
